surface.gap: Update the logic following the introduction of the new Identity class.
Note that at construction, the tokpass cannot be created with the token only and especially without the password. This part needs to be updated or completely redone through a SSH-like agent for instance.

Finally, the identity retrieved from the meta disciple should be validated but gap does not seem to have an authority which makes it impossible to perform such a verification. This point should also be improved.