infinit.Identity: Introduce the completely reworked Identity class which is neither Fileable nor Uniquable.
Note that unlike the previous version, an Identity instance cannot be encrypted or decrypted. An identity is already encrypted which is its internal code represents the user's keypair in its encrypted form. Should the user want to extract the keypair, he would simply need to call the decrypt(passphrase) method which returns the associated key pair.