surface.gap.PeerTranferMachine: Integrante UPNP-IGD support.
UPNP query takes time, so start it in backgronud at machine creation time
(which is when the station is created).

Use one UPNP object per transaction to be more robust to network condition
changes (until we have proper os-level detection, at which point it will
be possible to have one UPNP instance in State, and reinitialize it
when network condition changes).